Add reset for binary row on each number changes
Added a for loop so that all the columns in the binary row would be
reset to 0 each time the number input is changed. This is to ensure
that if the number is changed from a large number to a small number,
the columns are not incorrectly displaying values from the previous
input.

Commented out unneeded buttons and console.logs.